pyproject.toml
[tool.poetry]
name = "django-homework-29"
version = "0.1.0"
description = ""
authors = ["namazovanton <namazovanton@gmail.com>"]
readme = "README.md"
packages = [{include = "django_homework_29"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
